+++ pollux/lib/vc-jwt/src/test/scala/io/iohk/atala/pollux/vc/jwt/ES256KSignerTest.scala
@@ -16,7 +16,7 @@ import java.security.interfaces.ECPrivateKey
 class ES256KSignerTest extends AnyFunSuite with Matchers {
 
   test("ES256KSigner should use BouncyCastleProviderSingleton") {
-    val ecKey = ECKeyGenerator(Curve.SECP256K1).generate()
+    val ecKey = ECKeyGenerator(Curve.SECP256K1).provider(BouncyCastleProviderSingleton.getInstance()).generate()
     val signer = new ES256KSigner(ecKey.toPrivateKey).signer
     val provider = signer.getJCAContext.getProvider
     provider mustBe a[BouncyCastleProvider]